test(docx): move #4185 regression tests into test_backend_msword.py
